+++ social_sqlalchemy/storage.py
@@ -17,13 +17,27 @@ from sqlalchemy.ext.mutable import MutableDict
 from social_core.storage import UserMixin, AssociationMixin, NonceMixin, \
                                 CodeMixin, PartialMixin, BaseStorage
 
+class JSONPickler(object):
+    """JSON pickler wrapper around json lib since SQLAlchemy invokes
+    dumps with extra positional parameters"""
+
+    @classmethod
+    def dumps(cls, value, *args, **kwargs):
+        """Dumps the python value into a JSON string"""
+        return json.dumps(value)
+
+    @classmethod
+    def loads(cls, value):
+        """Parses the JSON string and returns the corresponding python value"""
+        return json.loads(value)
+
 
 # JSON type field
 class JSONType(PickleType):
     impl = Text
 
     def __init__(self, *args, **kwargs):
-        kwargs['pickler'] = json
+        kwargs['pickler'] = JSONPickler
         super(JSONType, self).__init__(*args, **kwargs)